pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.0.3",
    "python-dotenv>=1.1.1",
]
